ai-service/requirements.txt
fastapi
uvicorn
spacy
pdfminer.six
python-multipart
python-dotenv
pymongo
requests
